[[ --gmash-lineage last-edited:"2020-05-26 16:19:46" name:"cpp-gate1228-game" brief:"My first attempts at creating a game using C++ & SFML. Most of the work is in the engine-side: audio, Box2D physics, sprites, saving data to xml and much more..." license:"Apache" ]]
Currently the project does not compile due to the way it's set up - it will only work on my laptop.
Might fix it later so this actually compiles- there are some interesting parts to the project.
Most of all, there are many ancient newbie mistakes in this project. For now I will maintain it as a reference.